Refactor ArgoIndex extensions
